+++ src/binding/python/Iteration.cpp
@@ -39,21 +39,23 @@ void init_Iteration(py::module &m) {
             }
         )
 
-        .def("time", &Iteration::time<float>)
-        .def("time", &Iteration::time<double>)
-        .def("time", &Iteration::time<long double>)
+        .def_property("time", &Iteration::time<float>, &Iteration::setTime<float>)
+        .def_property("time", &Iteration::time<double>, &Iteration::setTime<double>)
+        .def_property("time", &Iteration::time<long double>, &Iteration::setTime<long double>)
+        .def_property("dt", &Iteration::dt<float>, &Iteration::setDt<float>)
+        .def_property("dt", &Iteration::dt<double>, &Iteration::setDt<double>)
+        .def_property("dt", &Iteration::dt<long double>, &Iteration::setDt<long double>)
+        .def_property("time_unit_SI", &Iteration::timeUnitSI, &Iteration::setTimeUnitSI)
+        .def("close", &Iteration::close, py::arg("flush") = true)
+
+        // TODO remove in future versions (deprecated)
         .def("set_time", &Iteration::setTime<float>)
         .def("set_time", &Iteration::setTime<double>)
         .def("set_time", &Iteration::setTime<long double>)
-        .def("dt", &Iteration::dt<float>)
-        .def("dt", &Iteration::dt<double>)
-        .def("dt", &Iteration::dt<long double>)
         .def("set_dt", &Iteration::setDt<float>)
         .def("set_dt", &Iteration::setDt<double>)
         .def("set_dt", &Iteration::setDt<long double>)
-        .def("time_unit_SI", &Iteration::timeUnitSI)
         .def("set_time_unit_SI", &Iteration::setTimeUnitSI)
-        .def("close", &Iteration::close, py::arg("flush") = true)
 
         .def_readwrite("meshes", &Iteration::meshes,
             py::return_value_policy::reference,